isp-framework/requirements-ai.txt
# AI-First Development and Testing Tools
# These tools support AI-generated code development and testing

# Property-based testing framework
hypothesis>=6.92.0
hypothesis[django]>=6.92.0

# API contract testing
schemathesis>=3.19.0
jsonschema>=4.19.0

# Performance regression detection  
pytest-benchmark>=4.0.0
pytest-memray>=1.4.0

# Parallel test execution
pytest-xdist>=3.3.1
pytest-forked>=1.6.0

# Enhanced test reporting
pytest-html>=4.1.1
pytest-json-report>=1.5.0
pytest-cov>=4.1.0

# Security scanning tools
bandit[toml]>=1.7.5
semgrep>=1.45.0
safety>=2.3.0
pip-audit>=2.6.1

# Code analysis for AI detection
ast-grep>=0.11.0
libcst>=1.1.0

# Structured logging for AI monitoring
structlog>=23.1.0
rich>=13.5.0

# HTTP testing for contract validation  
requests>=2.31.0
responses>=0.23.3
httpx>=0.24.0

# Data generation for property testing
faker>=20.1.0
factory-boy>=3.3.0

# Performance monitoring
psutil>=5.9.0
memory-profiler>=0.61.0

# AI safety and monitoring tools
opentelemetry-api>=1.20.0
opentelemetry-sdk>=1.20.0
opentelemetry-instrumentation-requests>=0.41b0

# Development utilities
watchdog>=3.0.0
python-dotenv>=1.0.0

# Database testing utilities (for contract tests)
pytest-postgresql>=5.0.0
pytest-redis>=3.0.0

# Async testing support
pytest-asyncio>=0.21.0
aioresponses>=0.7.4

# Mock and stubbing for behavior tests
respx>=0.20.2
pytest-mock>=3.11.0

# Test data validation
cerberus>=1.3.4
pydantic>=2.4.0

# Configuration testing
dynaconf>=3.2.0
python-decouple>=3.8